requirements.txt
streamlit
langchain
openai
python-dotenv
google-generativeai
langchain-google-genai
audio-recorder-streamlit
openai  
pydub